Changed login functionality so it is shared between distributors and business
