+++ config/routes.rb
@@ -11,14 +11,11 @@ Rails.application.routes.draw do
       get '/business_items/:id', to: 'business_items#show'
       post '/business_items', to: 'business_items#create'
       get '/cocktails/:id/', to: 'cocktails#show'
+      post '/login', to: 'login#create'
       namespace :businesses do
         post '/register', to: 'registration#create'
-        post '/login', to: 'login#create'
         get '/account', to: 'account#show'
         get '/items_popularity', to: 'items_popularity#index'
-      end
-      namespace :distributors do
-        post '/login', to: 'login#create'
       end
     end
   end
